Module aws_sdk_kafka::model

Modules

broker_ebs_volume_info

See BrokerEbsVolumeInfo

broker_logs

See BrokerLogs

broker_node_group_info

See BrokerNodeGroupInfo

broker_node_info

See BrokerNodeInfo

broker_software_info

See BrokerSoftwareInfo

client_authentication

See ClientAuthentication

cloud_watch_logs

See CloudWatchLogs

cluster_info

See ClusterInfo

cluster_operation_info

See ClusterOperationInfo

cluster_operation_step

See ClusterOperationStep

cluster_operation_step_info

See ClusterOperationStepInfo

compatible_kafka_version

See CompatibleKafkaVersion

configuration

See Configuration

configuration_info

See ConfigurationInfo

configuration_revision

See ConfigurationRevision

ebs_storage_info

See EbsStorageInfo

encryption_at_rest

See EncryptionAtRest

encryption_in_transit

See EncryptionInTransit

encryption_info

See EncryptionInfo

error_info

See ErrorInfo

firehose

See Firehose

iam

See Iam

jmx_exporter

See JmxExporter

jmx_exporter_info

See JmxExporterInfo

kafka_version

See KafkaVersion

logging_info

See LoggingInfo

mutable_cluster_info

See MutableClusterInfo

node_exporter

See NodeExporter

node_exporter_info

See NodeExporterInfo

node_info

See NodeInfo

open_monitoring

See OpenMonitoring

open_monitoring_info

See OpenMonitoringInfo

prometheus

See Prometheus

prometheus_info

See PrometheusInfo

s3

See S3

sasl

See Sasl

scram

See Scram

state_info

See StateInfo

storage_info

See StorageInfo

tls

See Tls

unauthenticated

See Unauthenticated

unprocessed_scram_secret

See UnprocessedScramSecret

zookeeper_node_info

See ZookeeperNodeInfo

Structs

BrokerEbsVolumeInfo

Specifies the EBS volume upgrade information. The broker identifier must be set to the keyword ALL. This means the changes apply to all the brokers in the cluster.

BrokerLogs

BrokerNodeGroupInfo

Describes the setup to be used for Kafka broker nodes in the cluster.

BrokerNodeInfo

BrokerNodeInfo

BrokerSoftwareInfo

Information about the current software installed on the cluster.

ClientAuthentication

Includes all client authentication information.

CloudWatchLogs

ClusterInfo

Returns information about a cluster.

ClusterOperationInfo

Returns information about a cluster operation.

ClusterOperationStep

Step taken during a cluster operation.

ClusterOperationStepInfo

State information about the operation step.

CompatibleKafkaVersion

Contains source Kafka versions and compatible target Kafka versions.

Configuration

Represents an MSK Configuration.

ConfigurationInfo

Specifies the configuration to use for the brokers.

ConfigurationRevision

Describes a configuration revision.

EbsStorageInfo

Contains information about the EBS storage volumes attached to Kafka broker nodes.

EncryptionAtRest

The data-volume encryption details.

EncryptionInTransit

The settings for encrypting data in transit.

EncryptionInfo

Includes encryption-related information, such as the AWS KMS key used for encrypting data at rest and whether you want MSK to encrypt your data in transit.

ErrorInfo

Returns information about an error state of the cluster.

Firehose

Iam

Details for IAM access control.

JmxExporter

Indicates whether you want to enable or disable the JMX Exporter.

JmxExporterInfo

Indicates whether you want to enable or disable the JMX Exporter.

KafkaVersion

LoggingInfo

MutableClusterInfo

Information about cluster attributes that can be updated via update APIs.

NodeExporter

Indicates whether you want to enable or disable the Node Exporter.

NodeExporterInfo

Indicates whether you want to enable or disable the Node Exporter.

NodeInfo

The node information object.

OpenMonitoring

JMX and Node monitoring for the MSK cluster.

OpenMonitoringInfo

JMX and Node monitoring for the MSK cluster.

Prometheus

Prometheus settings.

PrometheusInfo

Prometheus settings.

S3

Sasl

Details for client authentication using SASL.

Scram

Details for SASL/SCRAM client authentication.

StateInfo

StorageInfo

Contains information about storage volumes attached to MSK broker nodes.

Tls

Details for client authentication using TLS.

Unauthenticated

UnprocessedScramSecret

Error info for scram secret associate/disassociate failure.

ZookeeperNodeInfo

Zookeeper node information.

Enums

BrokerAzDistribution

The distribution of broker nodes across Availability Zones. This is an optional parameter. If you don't specify it, Amazon MSK gives it the value DEFAULT. You can also explicitly set this parameter to the value DEFAULT. No other values are currently allowed.

ClientBroker

Client-broker encryption in transit setting.

ClusterState

The state of a Kafka cluster.

ConfigurationState

The state of a configuration.

EnhancedMonitoring

Specifies which metrics are gathered for the MSK cluster. This property has the following possible values: DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION. For a list of the metrics associated with each of these levels of monitoring, see Monitoring.

KafkaVersionStatus

NodeType

The broker or Zookeeper node.