Module aws_sdk_kafka::operation

Structs

BatchAssociateScramSecret

Associates one or more Scram Secrets with an Amazon MSK cluster.

BatchDisassociateScramSecret

Disassociates one or more Scram Secrets from an Amazon MSK cluster.

CreateCluster

Creates a new MSK cluster.

CreateConfiguration

Creates a new MSK configuration.

DeleteCluster

Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request.

DeleteConfiguration

Deletes an MSK Configuration.

DescribeCluster

Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.

DescribeClusterOperation

Returns a description of the cluster operation specified by the ARN.

DescribeConfiguration

Returns a description of this MSK configuration.

DescribeConfigurationRevision

Returns a description of this revision of the configuration.

GetBootstrapBrokers

A list of brokers that a client application can use to bootstrap.

GetCompatibleKafkaVersions

Gets the Apache Kafka versions to which you can update the MSK cluster.

ListClusterOperations

Returns a list of all the operations that have been performed on the specified MSK cluster.

ListClusters

Returns a list of all the MSK clusters in the current Region.

ListConfigurationRevisions

Returns a list of all the MSK configurations in this Region.

ListConfigurations

Returns a list of all the MSK configurations in this Region.

ListKafkaVersions

Returns a list of Kafka versions.

ListNodes

Returns a list of the broker nodes in the cluster.

ListScramSecrets

Returns a list of the Scram Secrets associated with an Amazon MSK cluster.

ListTagsForResource

Returns a list of the tags associated with the specified resource.

RebootBroker

Reboots brokers.

TagResource

Adds tags to the specified MSK resource.

UntagResource

Removes the tags associated with the keys that are provided in the query.

UpdateBrokerCount

Updates the number of broker nodes in the cluster.

UpdateBrokerStorage

Updates the EBS storage associated with MSK brokers.

UpdateBrokerType

Updates EC2 instance type.

UpdateClusterConfiguration

Updates the cluster with the configuration that is specified in the request body.

UpdateClusterKafkaVersion

Updates the Apache Kafka version for the cluster.

UpdateConfiguration

Updates an MSK configuration.

UpdateMonitoring

Updates the monitoring settings for the cluster. You can use this operation to specify which Apache Kafka metrics you want Amazon MSK to send to Amazon CloudWatch. You can also specify settings for open monitoring with Prometheus.

UpdateSecurity

Updates the security settings for the cluster. You can use this operation to specify encryption and authentication on existing clusters.